Implement the StudentAccount Class
for my part of this project, I implemented the Student Account Class. According to the specifications the class should 1. All getters should take a Session parameter and throw an ExpiredSessionException if the session is expired or an AccessViolationException if the session owner doesn't match the account owner and the session owner is not an Administrator or Faculty. 2. All setters should take a Session parameter and throw an ExpiredSessionException if the session is expired or an AccessViolationException if the session owner doesn't match the account owner and the session owner is not an Administrator or Faculty. Also, setters for loginName and academicHistory should be restricted and throw an AccessViolationException if an attempt is made to modify them by anyone other than an Administrator or Faculty.(I also added a bunch of comments to better explain the structure and intention of the codes to stay true to the lessons learnt in class).